heater-shaker: scripts: modularize
Pull a bunch of stuff from speed_stability to new file test_utils for reuse
